+++ x/cronos/keeper/evm_hooks.go
@@ -1,12 +1,13 @@
 package keeper
 
 import (
+	"math/big"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/crypto-org-chain/cronos/x/cronos/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-
-	"github.com/crypto-org-chain/cronos/x/cronos/types"
 )
 
 // LogProcessEvmHook is an evm hook that convert specific contract logs into native module calls
@@ -26,6 +27,7 @@ func NewLogProcessEvmHook(handlers ...types.EvmLogHandler) *LogProcessEvmHook {
 
 // PostTxProcessing implements EvmHook interface
 func (h LogProcessEvmHook) PostTxProcessing(ctx sdk.Context, msg core.Message, receipt *ethtypes.Receipt) error {
+	addLogToReceiptFunc := newFuncAddLogToReceipt(receipt)
 	for _, log := range receipt.Logs {
 		if len(log.Topics) == 0 {
 			continue
@@ -34,10 +36,37 @@ func (h LogProcessEvmHook) PostTxProcessing(ctx sdk.Context, msg core.Message, r
 		if !ok {
 			continue
 		}
-		err := handler.Handle(ctx, log.Address, log.Data)
+		err := handler.Handle(ctx, log.Address, log.Data, addLogToReceiptFunc)
 		if err != nil {
 			return err
 		}
 	}
 	return nil
+}
+
+// newFuncAddLogToReceipt return a function to add additional logs to the receipt
+func newFuncAddLogToReceipt(receipt *ethtypes.Receipt) func(logSig common.Hash, logData []byte) {
+	return func(logSig common.Hash, logData []byte) {
+		if receipt.BlockNumber == nil {
+			return
+		}
+		newLog := &ethtypes.Log{
+			Address:     receipt.ContractAddress,
+			Topics:      []common.Hash{logSig},
+			Data:        logData,
+			BlockNumber: receipt.BlockNumber.Uint64(),
+			TxHash:      receipt.TxHash,
+			TxIndex:     receipt.TransactionIndex,
+			BlockHash:   receipt.BlockHash,
+			Index:       uint(len(receipt.Logs) + 1),
+			Removed:     false,
+		}
+
+		// Compute block bloom filter and set to the receipt
+		bloom := receipt.Bloom.Big()
+		bloom.Or(bloom, big.NewInt(0).SetBytes(ethtypes.LogsBloom([]*ethtypes.Log{newLog})))
+		receipt.Bloom = ethtypes.BytesToBloom(bloom.Bytes())
+
+		receipt.Logs = append(receipt.Logs, newLog)
+	}
 }